requirements.txt
scikit-learn
numpy
nltk
pyfunctional==0.7.1
pycountry
